Successfully connect to the database
