+++ backend/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "postman.settings.dotenv-detection-notification-visibility": false
+}
