+++ backend/requirements.txt
@@ -5,3 +5,4 @@ python-dotenv
 requests
 google-genai
 sqlalchemy
+greenlet
