+++ backend/routers/__init__.py
@@ -1,0 +1,4 @@
+from .chat import router as chat_router
+from .problems import router as problems_router
+
+__all__ = ["chat_router", "problems_router"]
